requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.3
pypdf>=3.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0